requirements.txt
  
Django==2.2.3    
Scrapy==1.6.0
    
scrapy-djangoitem==1.1.1    
scrapyd==1.2.1
python-scrapyd-api==2.1.2

django-crispy-forms==1.7.2    
django-filter==2.2.0    
django-widget-tweaks==1.4.5
bootstrap4==0.1.0

 